# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: arc — Personal Brand Strategy API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List all users with a summary of their data (admin only)
      responses:
        "200":
          description: All users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUserSummary"
        "403":
          description: Forbidden
  /admin/users/{clientId}:
    get:
      operationId: getAdminUser
      tags: [admin]
      summary: Get a single user's full profile, narrative, audit, posts, and ideas (admin only)
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserDetail"
        "403":
          description: Forbidden
        "404":
          description: Not found
...
components:
  schemas:
...
    ClientProfile:
      type: object
      required: [id, fullName, location, headline, currentRole, company, industry, yearsExperience, achievements, goals, bio, placeOfBirth, earlyLife, schooling, university, professionalJourney, signatureAchievements, awards, quantifiableResults, audienceImpact, passions, beliefs, frustrations, desiredChange, positioning, primaryAudience, secondaryAudience, geographyCulture, brandValues, nonNegotiables, personalityTone, desiredFeeling, thesis, coreBeliefs, signatureFrameworks, extractedInfo, onboardingComplete, onboardingStep, foundationConsolidatedAck, createdAt, updatedAt]
      properties:
        id:
          type: integer
        fullName:
          type: string
        location:
          type: string
        headline:
          type: string
        currentRole:
          type: string
        company:
          type: string
        industry:
          type: string
        yearsExperience:
          type: integer
        achievements:
          type: array
          items:
            type: string
        goals:
          type: string
        bio:
          type: string
        dateOfBirth:
          type: ["string", "null"]
        placeOfBirth:
          type: string
        earlyLife:
          type: string
        schooling:
          type: string
        university:
          type: string
        professionalJourney:
          type: string
        signatureAchievements:
          type: string
        awards:
          type: string
        quantifiableResults:
          type: string
        audienceImpact:
          type: string
        passions:
          type: string
        beliefs:
          type: string
        frustrations:
          type: string
        desiredChange:
          type: string
        positioning:
          type: string
        primaryAudience:
          type: string
        secondaryAudience:
          type: string
        geographyCulture:
          type: string
        brandValues:
          type: string
        nonNegotiables:
          type: string
        personalityTone:
          type: string
        desiredFeeling:
          type: string
        thesis:
          type: string
        coreBeliefs:
          type: string
        signatureFrameworks:
          type: string
        extractedInfo:
          type: string
        website:
          type: ["string", "null"]
        newsletter:
          type: ["string", "null"]
        linkedinUrl:
          type: ["string", "null"]
        twitterUrl:
          type: ["string", "null"]
        instagramUrl:
          type: ["string", "null"]
        youtubeUrl:
          type: ["string", "null"]
        onboardingComplete:
          type: boolean
        onboardingStep:
          type: integer
        foundationConsolidatedAck:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
...
    SeoFinding:
      type: object
      required: [title, url, type, snippet]
      properties:
        title:
          type: string
        url:
          type: string
        type:
          type: string
          enum: [owned, social, press, directory, other]
        snippet:
          type: string
    SeoFindings:
      type: object
      required: [resultCount, results, ownedPresence, summary]
      properties:
        resultCount:
          type: integer
        results:
          type: array
          items:
            $ref: "#/components/schemas/SeoFinding"
        ownedPresence:
          type: boolean
        summary:
          type: string
    GeoModelResult:
      type: object
      required: [model, label, mentioned, accuracy, response, notes]
      properties:
        model:
          type: string
        label:
          type: string
        mentioned:
          type: boolean
        accuracy:
          type: string
          enum: [accurate, partial, none, incorrect]
        response:
          type: string
        notes:
          type: string
    GeoSource:
      type: object
      required: [title, url]
      properties:
        title:
          type: string
        url:
          type: string
    GeoFindings:
      type: object
      required: [models, summary]
      properties:
        models:
          type: array
          items:
            $ref: "#/components/schemas/GeoModelResult"
        summary:
          type: string
        sources:
          type: array
          items:
            $ref: "#/components/schemas/GeoSource"
    AuditResult:
      type: object
      required: [id, clientId, seoScore, geoScore, recommendations, status, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        seoScore:
          type: integer
        geoScore:
          type: integer
        seoFindings:
          oneOf:
            - $ref: "#/components/schemas/SeoFindings"
            - type: "null"
        geoFindings:
          oneOf:
            - $ref: "#/components/schemas/GeoFindings"
            - type: "null"
        recommendations:
          type: array
          items:
            type: string
        status:
          type: string
        createdAt:
          type: string
...
    NarrativeTheme:
      type: object
      required: [title, description]
      properties:
        title:
          type: string
        description:
          type: string
    PlatformRecommendation:
      type: object
      required: [platform, reason, priority]
      properties:
        platform:
          type: string
        reason:
          type: string
        priority:
          type: string
          enum: [high, medium, low]
    IndustryAnswer:
      type: object
      required: [question, answer]
      properties:
        question:
          type: string
        answer:
          type: string
...
    NarrativeProfile:
      type: object
      required: [id, clientId, industryAnswers, coreNarrative, pointOfView, themes, recommendedPlatforms, contentHooks, createdAt, updatedAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        industryAnswers:
          type: array
          items:
            $ref: "#/components/schemas/IndustryAnswer"
        coreNarrative:
          type: string
        pointOfView:
          type: string
        themes:
          type: array
          items:
            $ref: "#/components/schemas/NarrativeTheme"
        recommendedPlatforms:
          type: array
          items:
            $ref: "#/components/schemas/PlatformRecommendation"
        contentHooks:
          type: array
          items:
            type: string
        createdAt:
          type: string
        updatedAt:
          type: string
...
    Post:
      type: object
      required: [id, title, content, platform, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        platform:
          type: string
          enum: [linkedin, twitter, instagram, blog, other]
        status:
          type: string
          enum: [draft, scheduled, published]
        scheduledAt:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        handoffProvider:
          type: ["string", "null"]
        handoffAt:
          type: ["string", "null"]
        handoffRef:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
...
    Idea:
      type: object
      required: [id, title, notes, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        notes:
          type: string
        platform:
          type: ["string", "null"]
        createdAt:
          type: string
...
    AdminUserSummary:
      type: object
      required: [clientId, userId, name, onboardingComplete, auditComplete, narrativeComplete, postCount, ideaCount, createdAt, updatedAt]
      properties:
        clientId:
          type: integer
        userId:
          type: string
        email:
          type: ["string", "null"]
        name:
          type: string
        onboardingComplete:
          type: boolean
        seoScore:
          type: ["integer", "null"]
        geoScore:
          type: ["integer", "null"]
        auditComplete:
          type: boolean
        narrativeComplete:
          type: boolean
        postCount:
          type: integer
        ideaCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
    AdminUserDetail:
      type: object
      required: [clientId, userId, profile, posts, ideas]
      properties:
        clientId:
          type: integer
        userId:
          type: string
        email:
          type: ["string", "null"]
        profile:
          $ref: "#/components/schemas/ClientProfile"
        audit:
          oneOf:
            - $ref: "#/components/schemas/AuditResult"
            - type: "null"
        narrative:
          oneOf:
            - $ref: "#/components/schemas/NarrativeProfile"
            - type: "null"
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        ideas:
          type: array
          items:
            $ref: "#/components/schemas/Idea"
...